Simplify SwiftEnvironment just to help with GlobalEnvironment
