+++ Sources/SwiftEnvironment/InstanceResolver/InstanceResolver.swift
@@ -10,5 +10,4 @@ import SwiftUI
 
 public protocol InstanceResolver {
     func resolve<V>(for type: V.Type) -> V?
-    func assign(to view: any View, for keyPath: AnyKeyPath) -> any View
 }
