+++ Sources/SwiftEnvironment/InstanceResolver/SingletonInstanceResolver.swift
@@ -30,12 +30,4 @@ final class SingletonInstanceResolver<Value>: InstanceResolver {
         }
         return instance as? V
     }
-    
-    @inlinable func assign(to view: any View, for keyPath: AnyKeyPath) -> any View {
-        guard let writableKeyPath = keyPath as? WritableKeyPath<EnvironmentValues, Value>,
-              let value = resolve(for: Value.self) else {
-            return view
-        }
-        return view.environment(writableKeyPath, value)
-    }
 }
