+++ Sources/SwiftEnvironment/InstanceResolver/WeakInstanceResolver.swift
@@ -9,9 +9,9 @@ import Foundation
 import Chary
 import SwiftUI
 
-final class WeakInstanceResolver<Value: AnyObject>: InstanceResolver {
+final class WeakInstanceResolver<Value>: InstanceResolver {
     
-    private(set) weak var instance: Value?
+    private(set) weak var instance: AnyObject?
     private let resolver: () -> Value
     private let queue: DispatchQueue?
     
@@ -23,17 +23,9 @@ final class WeakInstanceResolver<Value: AnyObject>: InstanceResolver {
     @inlinable func resolve<V>(for type: V.Type) -> V? {
         guard let instance else {
             let newInstance = queue?.safeSync(execute: resolver) ?? resolver()
-            self.instance = newInstance
+            self.instance = newInstance as AnyObject
             return newInstance as? V
         }
         return instance as? V
-    }
-    
-    @inlinable func assign(to view: any View, for keyPath: AnyKeyPath) -> any View {
-        guard let writableKeyPath = keyPath as? WritableKeyPath<EnvironmentValues, Value>,
-              let value = resolve(for: Value.self) else {
-            return view
-        }
-        return view.environment(writableKeyPath, value)
     }
 }
