+++ Sources/SwiftEnvironment/PropertyWrapperDiscardable.swift
@@ -1,0 +1,31 @@
+//
+//  PropertyWrapperDiscardable.swift
+//  SwiftEnvironment
+//
+//  Created by Nayanda Haberty on 10/03/25.
+//
+
+import Foundation
+
+public protocol PropertyWrapperDiscardable {
+    func discardValueSet()
+    var projectedValue: PropertyWrapperDiscardableControl { get }
+}
+
+extension PropertyWrapperDiscardable {
+    public var projectedValue: PropertyWrapperDiscardableControl {
+        PropertyWrapperDiscardableControl(propertyWrapper: self)
+    }
+}
+
+public struct PropertyWrapperDiscardableControl {
+    private let propertyWrapper: PropertyWrapperDiscardable
+    
+    @usableFromInline init(propertyWrapper: PropertyWrapperDiscardable) {
+        self.propertyWrapper = propertyWrapper
+    }
+    
+    public func discardValueSet() {
+        propertyWrapper.discardValueSet()
+    }
+}
