--- Sources/SwiftEnvironment/Resolver/EnvironmentValuesResolverHost.swift
@@ -1,36 +1,0 @@
-//
-//  ViewEnvironment.swift
-//  SwiftEnvironment
-//
-//  Created by Nayanda Haberty on 5/11/24.
-//
-
-import Foundation
-
-protocol EnvironmentValuesResolverHost: EnvironmentValuesResolving {
-    var environmentValuesResolver: EnvironmentValuesResolving { get }
-}
-
-extension EnvironmentValuesResolverHost {
-    
-    @inlinable public func resolve<V>(_ keyPath: KeyPath<EnvironmentValues, V>) -> V {
-        environmentValuesResolver.resolve(keyPath)
-    }
-    
-    @discardableResult
-    @inlinable public func environment<V>(
-        _ keyPath: WritableKeyPath<EnvironmentValues, V>,
-        resolveOn queue: DispatchQueue?,
-        resolver: @escaping () -> V) -> Self {
-            environmentValuesResolver.environment(keyPath, resolveOn: queue, resolver: resolver)
-            return self
-        }
-    
-    @discardableResult
-    @inlinable public func environment<S, V>(
-        _ keyPath: WritableKeyPath<EnvironmentValues, V>,
-        use soureKeyPath: WritableKeyPath<EnvironmentValues, S>) -> Self {
-            environmentValuesResolver.environment(keyPath, use: soureKeyPath)
-            return self
-        }
-}
